Ballomatic9000.jl Documentation

Ballomatic9000.Domain — Type

The Domain representing the locations of the walls, which are at:

x = 0

y = 0

x = xmax

y = ymax

Ballomatic9000.calculate_time_from_coeffs — Method

calculate_time_from_coeffs(coeffs)

Coeffs of a polynomial of form ∑ᵢ₌₀ᴺ cᵢtⁱ

Return the smallest non-zero positive root corresonding the time of collision.

...

Arguments

coeffs: Vector-like container of numbers

...

Example

Ballomatic9000.collide! — Function

collide!(a::Ball,d::Domain,atol=sqrt(eps()),rtol=sqrt(eps()))

Collide a ball with the wall of the Domain

...

Arguments

a::Ball: ball that could collide with the domain wall

d::Domain: domain defining positions of the wall

atol=sqrt(eps()): absolute tolerance to check position of ball wrt walls

rtol=sqrt(eps()): as above but relative tolerance

...

Example

Ballomatic9000.collide! — Method

collide!(a::Ball,b::Ball)

Collide two balls off one another in an elastic collision that alters their velocities. No temporal advance is taken.

See https://en.wikipedia.org/wiki/Elastic_collision

...

Arguments

a::Ball:

b::Ball:

...

Example

Ballomatic9000.timeofcollision — Method

timeofcollision(a::Ball,force,d::Domain)

Calculate the time that the ball will collide with the walls of the domain

...

Arguments

a::Ball: the ball

force: the force acting on the ball

d::Domain: the domain in which the ball is bouncing

...

Example

Ballomatic9000.timeofcollision — Method

timeofcollision(a::Ball,b::Ball,forcea,forceb)

Calculate all the collisions that the two accelerating balls may undertake and return the soonest collision time.

Position at time t for ball a, x1' = x1 + v1t + a1/2 * t^2 Position at time t for ball b, x2' = x2 + v2t + a2/2 * t^2

Difference in positions at time t, (x1' - x2') = (x1 - x2) + (v1 - v2)t + (a1 - a2)/2 * t^2 Square of distance between balls at time t, (x1' - x2')^2 = ((x1 - x2) + (v1 - v2)t + (a1 - a2)/2 * t^2)^2 At the time of collision that squared distance is (r1 + r2)^2 = (x1' - x2')^2

import sympy x1, x2, v1, v2, a1, a2, t, r1, r2 = sympy.symbols('x1 x2 v1 v2 a1 a2 t r1 r2') expr = ((x1 - x2) + (v1 - v2)t + (a1 - a2)t2 / 2)2 - (r1 + r2)**2 sympy.Poly(expr, t).all_coeffs()

[a12/4 - a1*a2/2 + a22/4, a1v1 - a1v2 - a2v1 + a2v2, a1x1 - a1x2 - a2x1 + a2x2 + v12 - 2v1v2 + v22, 2v1x1 - 2v1x2 - 2v2x1 + 2v2x2, -r12 - 2r1r2 - r22 + x12 - 2x1x2 + x22] ...

Arguments

a::Ball: ball to collide

b::Ball: ball to collide

forcea: force on ball a

forceb: force on ball b

...

Example

Index

Ballomatic9000.Domain

Ballomatic9000.calculate_time_from_coeffs

Ballomatic9000.collide!

Ballomatic9000.collide!

Ballomatic9000.timeofcollision

Ballomatic9000.timeofcollision